Move GetEnumGpuIndicesText from D3d header to main.cpp
